refactored code and created web mod
