+++ src/web/mod.rs
@@ -1,0 +1,2 @@
+pub mod dom;
+pub mod canvas;
